+++ test/__tests__/backup.test.js
@@ -8,7 +8,7 @@ const supertest = require('supertest');
 const request = supertest(app);
 
 const masterCredentials = {
-  name: 'master',
+  username: 'master',
   password: 'master',
 };