+++ test/jest.config.js
@@ -3,4 +3,5 @@ module.exports = {
   rootDir: '../',
   moduleFileExtensions: ['ts', 'js', 'json'],
   testEnvironment: 'node',
+  testRunner: './node_modules/jest-jasmine2/build/index',
 };